Beyond dotfiles: adding CLI tools.
I added two small tools I developed, one that transforms Markdown documents into
nice PDFs and one that generates pronounceable passwords.

The goal of this repository is now to contain most of the things I'd need to get
up and running on a new machine. I decided to make it public, primarily to be
able to clone it anonymously.